feat: Add email dto
